исправил записи

+++ src/client_subjects/client_subjects.controller.ts
@@ -240,6 +240,29 @@ export class Client_subjectsController {
     }
   }
 
+  @Delete('deleteMany')
+  @ApiOperation({ summary: 'Удалить связи клиентов и предметов по ID' })
+  @ApiResponse({
+    status: 200,
+    description: 'Связи клиентов и предметов успешно удалены',
+    type: client_subjectsDto,
+  })
+  @ApiResponse({
+    status: 404,
+    description: 'Связи клиентов и предметов не найдены',
+  })
+  @ApiBearerAuth()
+  @UseGuards(JwtAuthGuard)
+  async deleteManyClient_subject(@Body('ids') ids: number[]): Promise<number> {
+    if (!ids || ids.length === 0) {
+      throw new BadRequestException(
+        'IDs связей клиентов и предметов не переданы',
+      );
+    }
+
+    return this.client_subjectsService.deleteManyClient_subjectsService(ids);
+  }
+
   @Delete(':id')
   @ApiOperation({ summary: 'Удалить связь клиентов и предметов по ID' })
   @ApiResponse({
@@ -265,6 +288,10 @@ export class Client_subjectsController {
       throw new BadRequestException('ID связи клиентов и предметов не передан');
     }
 
+    if (id) {
+      throw new BadRequestException('ID связи клиента и предмета некорректен');
+    }
+
     const result =
       await this.client_subjectsService.deleteClient_subjectService(Number(id));
 
@@ -273,28 +300,5 @@ export class Client_subjectsController {
     } else {
       throw new NotFoundException(`Связь клиентов и предметов не найдена`);
     }
-  }
-
-  @Delete('/deleteMany')
-  @ApiOperation({ summary: 'Удалить связи клиентов и предметов по ID' })
-  @ApiResponse({
-    status: 200,
-    description: 'Связи клиентов и предметов успешно удалены',
-    type: client_subjectsDto,
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Связи клиентов и предметов не найдены',
-  })
-  @ApiBearerAuth()
-  @UseGuards(JwtAuthGuard)
-  async deleteManyClient_subject(@Body('ids') ids: number[]): Promise<number> {
-    if (!ids || ids.length === 0) {
-      throw new BadRequestException(
-        'IDs связей клиентов и предметов не переданы',
-      );
-    }
-
-    return this.client_subjectsService.deleteManyClient_subjectsService(ids);
   }
 }
